02_c2d_messages/client/requirements.txt
azure-iot-device
python-dotenv